pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mofc-financials"
version = "0.1.0"
description = "Extract and analyze Mid-Ohio Food Collective IRS 990 financial data via OCR."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "PyMuPDF>=1.23",
    "pytesseract>=0.3",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "isort>=5.12",
    "mypy>=1.10",
    "interrogate>=1.7",
]
# Future: statistical analysis and visualization
analysis = [
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.13",
]

[project.scripts]
mofc-extract = "mofc_financials.data_extraction.extract_990:main"
mofc-extract-detail = "mofc_financials.data_extraction.extract_990_detail:main"
mofc-pipeline = "mofc_financials.data_extraction.validate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
known_first_party = ["mofc_financials"]
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not smoke'"
markers = [
    "smoke: slow OCR-based tests requiring Tesseract and real PDFs",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["fitz", "pytesseract"]
ignore_missing_imports = true

[tool.interrogate]
ignore-init-module = true
ignore-init-method = true
fail-under = 100
exclude = ["tests"]
verbose = 1
